+++ NEWS.md
@@ -1,3 +1,7 @@
+# mlr3fairness 0.3.1-9000
+* `fairness_ tensor` has a new argument `normalize` that controls normalization.
+  Possible values are `"all"` (global) `"group"` (by group) and `"none"` (no normalization).
+
 # mlr3fairness 0.3.1
 
 * Minor update to improve stability of unit tests and vignette building on CRAN.